==> 1.ETL steam_games.ipynb <==
#Steam Games
#Importo Librerias
import pandas as pd
import re

#Asigno al DataFrame el archivo json steam_games.json.gz
#df_sg=pd.read_json('/content/drive/MyDrive/PI_ML_OPS-PT/Dataset/steam_games.json.gz',compression='gzip',lines=True,encoding='utf-8')
df_sg=pd.read_json('./Dataset/steam_games.json.gz',compression='gzip',lines=True,encoding='utf-8')

#Elimino filas totalmente vacias
df_sg.dropna(how='all',inplace=True)

#Busqueda de Id nulos
df_sg[(df_sg['id'].isnull())]

#Busqueda de aplicacion con similar nombre de Id nulo
df_sg[(df_sg['app_name'] == 'Batman: Arkham City - Game of the Year Edition')]

#Eliminacion de id nulos
df_sg.dropna(subset=['id'],inplace=True)

#Duplicados
duplicados = df_sg['id'].value_counts() > 1
df_sg[df_sg['id'].isin(duplicados[duplicados].index)]

#Eliminación de Duplicados
df_sg.drop_duplicates(subset=['id'],inplace=True)

#Elimino genres vacios
df_sg.dropna(subset=['genres'],inplace=True)

#Búsqueda de nombres de aplicaciones vacias
df_sg[df_sg['app_name'].isnull()]

#Eliminacion de aplicaciones vacias
df_sg.dropna(subset=['app_name'],inplace=True)

#Eliminacion de columnas no necesarias
df_sg = df_sg.drop(columns=['publisher','title','url','reviews_url','early_access','specs','tags','price','developer'])

#Validacion de columnas nulas
df_sg.isnull().sum()

#Eliminacion de release_date nulas
df_sg.dropna(subset=['release_date'],inplace=True)

#Filtro año de fechas correctas
import re

def filtrar_dataframe_formato_fecha(df, nombre_columna):
    # Definir el patron  regex 'YYYY-MM-DD'
    patron_fecha = re.compile(r'^\d{4}-\d{2}-\d{2}$')

    # Filtrar filas con el correcto formato de fecha
    mask = df[nombre_columna].astype(str).str.match(patron_fecha)
    df_filtrado = df[mask]

    return df_filtrado

df_sg = filtrar_dataframe_formato_fecha(df_sg, 'release_date')

# Convierto id a entero
df_sg['id'] = df_sg['id'].astype('int')

# Convierto fecha a Año
df_sg['release_date'] = pd.to_datetime(df_sg['release_date'], errors='coerce')
df_sg['release_date'] = df_sg['release_date'].dt.year


# Mover la columna 'id' al principio, renombrarla por item_id y cambiar el tipo a int
df_sg.columns.get_loc('id')
nuevas_columnas = ['id'] + [columna for columna in df_sg.columns if columna != 'id']
df_sg = df_sg[nuevas_columnas]
df_sg.rename(columns={'id': 'item_id'}, inplace=True)
df_sg['item_id'] = df_sg['item_id'].astype('int')
# Mostrar las primeras filas del DataFrame con 'id' en la primera posición
df_sg.head()

#Desanidar genres y eliminar columna index
df_sg = df_sg.explode('genres').reset_index()
df_sg = df_sg.drop(columns="index")
df_sg.head()

#Obtener generos unicos
#df_generos_diferentes = df_sg['genres'].unique()
#df_generos_diferentes.sort()

# Guardo Dataframe en csv
#df_sg.to_csv('/content/drive/MyDrive/PI_ML_OPS-PT/Dataset/df_sg.csv',index=False)
df_sg.to_csv('./Dataset/df_sg.csv',index=False)

df_sg
