==> 2.ETL user_items.ipynb <==
#User Items
#Importo Librerias
import pandas as pd
import ast
import gzip

# User_Items
List_User_Items = []
#with gzip.open(u"/content/drive/MyDrive/PI_ML_OPS-PT/Dataset/users_items.json.gz", 'rt', encoding='UTF-8') as f:
with gzip.open(u"./Dataset/users_items.json.gz", 'rt', encoding='UTF-8') as f:
#with open("/content/drive/MyDrive/PI_ML_OPS-PT/Dataset/user_reviews.json.gz", encoding='UTF-8') as f:
    for line in f.readlines():
        List_User_Items.append(ast.literal_eval(line))

df_ui = pd.DataFrame(List_User_Items)
df_ui = df_ui.explode('items').reset_index()
df_ui = df_ui.drop(columns="index")
df_ui = pd.concat([df_ui, pd.json_normalize(df_ui['items'])], axis=1)
df_ui = df_ui.drop(columns='items')

##Busqueda de item_id nulos
df_ui[df_ui['item_id'].isna()]

# Elimino Vacios
df_ui.dropna(inplace=True)

#Valido los nulos
df_ui.isnull().sum()

# Elimino columnas no necesarias
df_ui.drop(columns=['items_count','steam_id','user_url','playtime_2weeks'],inplace=True)

# Seteo el indice
df_ui.reset_index(drop=True, inplace=True)

# Convierto id y playtime_forever a entero
df_ui['playtime_forever'] = df_ui['playtime_forever'].astype('int')
df_ui['item_id'] = df_ui['item_id'].astype('int')

# Muestro Dataframe
df_ui.head()

#Elimino los play_time forever(año) 0
df_ui.drop(df_ui[df_ui['playtime_forever'] == 0].index, inplace=True)
df_ui.reset_index(drop=True, inplace=True)

# Muestro Dataframe
df_ui.head()

# Guardo Dataframe en csv comprimido
# df_ui.to_csv('/content/drive/MyDrive/PI_ML_OPS-PT/Dataset/df_ui.csv.gz',index=False, compression='gzip')
df_ui.to_csv('./Dataset/df_ui.csv.gz',index=False, compression='gzip')

#Crear archivo base de genres
df_sgui=pd.merge(df_sg[['item_id','genres','release_date']],df_ui[['item_id','playtime_forever']],on='item_id',how='left')

#df_ug=pd.merge(df_sg[['genres','release_date']],df_ui[['playtime_forever']],on='item_id',how='left')
#df_sgui
#df_ug[df_ug['item_id']==10]

# hago un merge teniendo todos los usuarios
#user_rev_item = pd.merge(user_items_proc[['user_id','item_id','playtime_forever']],user_review[['user_id','recommend']],on = 'user_id',how='left')
# Me quedo solo con los que estan nulos en reviews
#user_rev_item = user_rev_item[user_rev_item['recommend'].isnull()]
# Hago un agrupamiento por usuario y saco el id del que tiene mas minuntos.
#max_playtime_items = user_rev_item.groupby('user_id')['playtime_forever'].idxmax()
#result_df = user_rev_item.loc[max_playtime_items, ['user_id', 'item_id']]
#result_df.to_csv('../datasets/user_item_max_horas.csv',index=False) # Exporto para ser usado por la funcion

 # hago un merge teniendo todos los usuarios
user_rev_item = pd.merge(df_ui[['user_id','item_id','playtime_forever']],user_review[['user_id','recommend']],on = 'user_id',how='left')
# Me quedo solo con los que estan nulos en reviews
user_rev_item = user_rev_item[user_rev_item['recommend'].isnull()]
# Hago un agrupamiento por usuario y saco el id del que tiene mas minuntos.
max_playtime_items = user_rev_item.groupby('user_id')['playtime_forever'].idxmax()
result_df = user_rev_item.loc[max_playtime_items, ['user_id', 'item_id']]
result_df.to_csv('../datasets/user_item_max_horas.csv',index=False) # Exporto para ser usado por la funcion
